requirements.txt
asgiref==3.7.2
captcha==0.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.60.1
django-axes==6.3.0
django-countries==7.5.1
django-recaptcha==4.0.0
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0